exam/src/requirements.txt
pulsar-client==3.3.0
faker==20.1.0
hdfs==2.7.3